+++ recipes/windows.rb
@@ -1,8 +1,8 @@
 #
-# Cookbook Name:: autopatch-nativex
+# Cookbook Name:: autopatch-simplyadrian
 # Recipe:: windows
 #
-# Copyright (C) 2015 NativeX
+# Copyright (C) 2015 simplyadrian
 #
 # All rights reserved - Do Not Redistribute
 #
@@ -11,18 +11,18 @@
 
 template 'AutoPatch PowerShell Script' do
   source 'auto-patch.ps1.erb'
-  path "#{node['autopatch-nativex']['working_dir']}\\auto-patch.ps1"
-  action :delete if node['autopatch-nativex']['disable']
+  path "#{node['autopatch-simplyadrian']['working_dir']}\\auto-patch.ps1"
+  action :delete if node['autopatch-simplyadrian']['disable']
 end
 
 windows_task 'autopatch' do
-  user node['autopatch-nativex']['task_username']
-  frequency node['autopatch-nativex']['task_frequency']
-  frequency_modifier node['autopatch-nativex']['task_frequency_modifier']
-  day node['autopatch-nativex']['task_days']
-  months node['autopatch-nativex']['task_months']
-  start_time node['autopatch-nativex']['task_start_time']
+  user node['autopatch-simplyadrian']['task_username']
+  frequency node['autopatch-simplyadrian']['task_frequency']
+  frequency_modifier node['autopatch-simplyadrian']['task_frequency_modifier']
+  day node['autopatch-simplyadrian']['task_days']
+  months node['autopatch-simplyadrian']['task_months']
+  start_time node['autopatch-simplyadrian']['task_start_time']
   cwd 'C:'
-  command "PowerShell -ExecutionPolicy Bypass -Command \"#{node['autopatch-nativex']['working_dir']}\\auto-patch.ps1\"" 
-  action :delete if node['autopatch-nativex']['disable']
+  command "PowerShell -ExecutionPolicy Bypass -Command \"#{node['autopatch-simplyadrian']['working_dir']}\\auto-patch.ps1\"" 
+  action :delete if node['autopatch-simplyadrian']['disable']
 end